@@ -1,17 +1,20 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
-use App\Models\ExchangeStudent;
 use App\Models\Person;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(ExchangeStudent::class, function (Faker $faker) {
-    return [
-        'id_user' => factory(Person::class)->create()->id_user,
-        'id_country' => $faker->numberBetween(1, 50),
-        'id_accommodation' => $faker->numberBetween(1, 10),
-        'id_faculty' => $faker->numberBetween(1, 7),
-        'about' => $faker->sentence,
-    ];
-});
+class ExchangeStudentFactory extends Factory
+{
+    public function definition()
+    {
+        return [
+            'id_user' => Person::factory()->create()->id_user,
+            'id_country' => $this->faker->numberBetween(1, 50),
+            'id_accommodation' => $this->faker->numberBetween(1, 10),
+            'id_faculty' => $this->faker->numberBetween(1, 7),
+            'about' => $this->faker->sentence,
+        ];
+    }
+}
