@@ -3,6 +3,7 @@
 namespace App\Http\Middleware;
 
 use Closure;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class RedirectIfAuthenticated
@@ -11,25 +12,31 @@ class RedirectIfAuthenticated
      * Handle an incoming request.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \Closure  $next
-     * @param  string|null  $guard
-     * @return mixed
+     * @param  \Closure(\Illuminate\Http\Request): (\Illuminate\Http\Response|\Illuminate\Http\RedirectResponse)  $next
+     * @param  string|null  ...$guards
+     * @return \Illuminate\Http\Response|\Illuminate\Http\RedirectResponse
      */
-    public function handle($request, Closure $next, $guard = null)
+    public function handle(Request $request, Closure $next, ...$guards)
     {
-        if (Auth::guard($guard)->check()) {
-            $user = Auth::user();
+        $guards = empty($guards) ? [null] : $guards;
 
-            if ($guard === 'tandem') {
-                return redirect(route('tandem.main'));
-            }
+        foreach ($guards as $guard) {
+            if (Auth::guard($guard)->check()) {
+                $user = Auth::user();
 
-            if ($user->isPartak()) {
-                return redirect('/partak');
-            } else if ($user->isBuddy()) {
-                return redirect('/muj-buddy');
-            } else {
-                return redirect('/user/verify');
+                if ($guard === 'tandem') {
+                    return redirect(route('tandem.main'));
+                }
+
+                if ($user->isPartak()) {
+                    return redirect('/partak');
+                } else {
+                    if ($user->isBuddy()) {
+                        return redirect('/muj-buddy');
+                    } else {
+                        return redirect('/user/verify');
+                    }
+                }
             }
         }
 
