@@ -2,10 +2,10 @@
 
 namespace App\Http\Resources;
 
-use Illuminate\Http\Resources\Json\Resource;
+use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Support\Collection;
 
-class AutocompleteItemResource extends Resource
+class AutocompleteItemResource extends JsonResource
 {
     /**
      * Transform the resource into an array.
