@@ -1,12 +1,15 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
-use App\Models\AlumniNewsletter;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(AlumniNewsletter::class, function (Faker $faker) {
-    return [
-        //
-    ];
-});
+class AlumniNewsletterFactory extends Factory
+{
+    public function definition()
+    {
+        return [
+            //
+        ];
+    }
+}
