@@ -14,12 +14,13 @@ return [
     */
 
     'accepted' => ':attribute musí být přijat.',
+    'accepted_if' => ':attribute musí být přijat, pokud :other je :value.',
     'active_url' => ':attribute není platnou URL adresou.',
     'after'  => ':attribute musí být datum po :date.',
     'after_or_equal' => ':attribute musí být datum nejdříve :date.',
-    'alpha' => ':attribute může obsahovat pouze písmena.',
-    'alpha_dash' => ':attribute může obsahovat pouze písmena, číslice, pomlčky a podtržítka. České znaky (á, é, í, ó, ú, ů, ž, š, č, ř, ď, ť, ň) nejsou podporovány.',
-    'alpha_num' => ':attribute může obsahovat pouze písmena a číslice.',
+    'alpha' => ':attribute musí obsahovat pouze písmena.',
+    'alpha_dash' => ':attribute musí obsahovat pouze písmena, číslice, pomlčky a podtržítka. České znaky (á, é, í, ó, ú, ů, ž, š, č, ř, ď, ť, ň) nejsou podporovány.',
+    'alpha_num' => ':attribute musí obsahovat pouze písmena a číslice.',
     'array' => ':attribute musí být pole.',
     'before' => ':attribute musí být datum před :date.',
     'before_or_equal' => ':attribute musí být datum nejpozději :date.',
@@ -31,9 +32,12 @@ return [
     ],
     'boolean' => ':attribute musí být true nebo false.',
     'confirmed' => ':attribute nebylo potvrzeno.',
+    'current_password' => 'Zadané heslo není správné.',
     'date' => ':attribute musí být platné datum.',
     'date_equals' => ':attribute musí být datum shodné s :date.',
     'date_format' => ':attribute není platný formát data podle :format.',
+    'declined' => ':attribute musí být odmítnut.',
+    'declined_if' => ':attribute musí být odmítnut, pokud :other je :value.',
     'different' => ':attribute a :other se musí lišit.',
     'digits' => ':attribute musí být :digits pozic dlouhé.',
     'digits_between' => ':attribute musí být dlouhé nejméně :min a nejvíce :max pozic.',
@@ -41,6 +45,7 @@ return [
     'distinct' => ':attribute má duplicitní hodnotu.',
     'email' => ':attribute nemá platný formát.',
     'ends_with' => ':attribute musí končit některým z následujících řetězců: :values.',
+    'enum' => 'Zvolený :attribute není platný.',
     'exists' => 'Zvolená hodnota pro :attribute není platná.',
     'file' => ':attribute musí být soubor.',
     'filled' => ':attribute musí být vyplněno.',
@@ -76,6 +81,7 @@ return [
         'string' => ':attribute musí být dlouhý nejvýše :value znaků.',
         'array' => ':attribute mus obsahovat nejvýše :value položek.',
     ],
+    'mac_address' => 'The :attribute must be a valid MAC address.',
     'max' => [
         'numeric' => ':attribute musí být nižší než :max.',
         'file' => ':attribute musí být menší než :max Kilobytů.',
@@ -90,13 +96,19 @@ return [
         'string' => ':attribute musí obsahovat alespoň :min znaků.',
         'array' => ':attribute musí obsahovat alespoň :min prvků.',
     ],
+    'multiple_of' => ':attribute mus být násobkem :value.',
     'not_in' => 'Zvolená hodnota pro :attribute je neplatná.',
     'not_regex' => ':attribute má neplatný formát.',
     'numeric' => ':attribute musí být číslo.',
     // 'password' => 'Heslo je nesprávné.',
     'present' => ':attribute musí být vyplněno.',
+    'prohibited' => 'Pole :attribute není povoleno.',
+    'prohibited_if' => 'Pole :attribute není povoleno, pokud :other je :value.',
+    'prohibited_unless' => 'Pole :attribute není povoleno, ledaže :values obsahuje :other.',
+    'prohibits' => 'Pole :attribute vylučuje přítomnost pole :other.',
     'regex' => ':attribute nemá správný formát.',
     'required' => ':attribute musí být vyplněno.',
+    'required_array_keys' => 'Pole :attribute musí obsahovat následující záznamy: :values.',
     'required_if' => ':attribute musí být vyplněno pokud :other je :value.',
     'required_unless' => ':attribute musí být vyplněno dokud :other je v :values.',
     'required_with' => ':attribute musí být vyplněno pokud :values je vyplněno.',
@@ -115,7 +127,7 @@ return [
     'timezone' => ':attribute musí být platná časová zóna.',
     'unique' => ':attribute musí být unikátní.',
     'uploaded' => 'Nahrávání :attribute se nezdařilo.',
-    'url' => 'Formát :attribute je neplatný.',
+    'url' => ':attribute musí být platná URL.',
     'uuid' => ':attribute musí být platné UUID.',
 
     /*
